refactor: Move UWP elevation code to Toolkit
